Removes old QKD files.
